+++ test/support/domain.ex
@@ -16,6 +16,7 @@ defmodule AshWorkflowTest.Domain do
     resource AshWorkflowTest.SharedConditionalWorkflow
     resource AshWorkflowTest.AcceptWorkflow
     resource AshWorkflowTest.InitialFlagWorkflow
+    resource AshWorkflowTest.DerivedTerminalWorkflow
     resource AshWorkflowTest.RepeatingTimeoutWorkflow
     resource AshWorkflowTest.CustomQueueWorkflow
     resource AshWorkflowTest.FieldTimeoutWorkflow